Extract parser into its own class.
Add spec coverage.
